+++ docs/fr_FR/Changelog.rst
@@ -1,6 +1,8 @@
 ChangeLog
 =========
 
+- Interne: Parser: Erreur 'msgToLQICollector' masquée (2546).
+
 230228-BETA-1
 -------------